Move from embedmodel to embedding model
